Feature/Details
Detalles de las peliculas y descripcion del programa V1.0
