+++ package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "@tailwindcss/postcss": "^4.1.8",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-router-dom": "^7.6.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.25.0",
