+++ src/components/MovieCard/MovieCardStyle.module.css
@@ -1,4 +1,3 @@
-/* MovieCard.module.css */
 .card {
   border: 6px solid #898c99;
   border-radius: 8px;
