+++ src/index.css
@@ -26,5 +26,5 @@ body {
 
 h1 {
   text-align: center;
-  color: white;
+  color: black;
 }
